requirements.txt
Flask<1
Pillow
python-dateutil
requests
numpy
python-resize-image
opencv-python
pymongo